requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
gradio>=4.0.0
opencv-python-headless>=4.7.0
numpy>=1.24.0
